+++ daft-derive/tests/integration/main.rs
@@ -1,0 +1,7 @@
+//! Integration tests for daft-derive.
+//!
+//! Structuring integration tests this way results in a single binary so new
+//! test modules don't have to build a separate binary.
+
+mod basic;
+mod generics;
